+++ Formula/millennium-helpers-bin.rb
@@ -11,22 +11,22 @@ class MillenniumHelpersBin < Formula
 
   on_macos do
     on_arm do
-      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.1/millennium-helpers-v3.0.1-darwin-arm64.tar.gz"
+      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.2/millennium-helpers-v3.0.2-darwin-arm64.tar.gz"
       sha256 "e005028f78d78e1fd437bf049b07048d5c530c0b2fd04161f972857966401a64"
     end
     on_intel do
-      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.1/millennium-helpers-v3.0.1-darwin-amd64.tar.gz"
+      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.2/millennium-helpers-v3.0.2-darwin-amd64.tar.gz"
       sha256 "03a2cfdab469d9e0bd60b62f34e2adae1985ac4301eb0969d39bf8e5cdbaf2b4"
     end
   end
 
   on_linux do
     on_arm do
-      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.1/millennium-helpers-v3.0.1-linux-arm64.tar.gz"
+      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.2/millennium-helpers-v3.0.2-linux-arm64.tar.gz"
       sha256 "ee89e8cf39839cd631c14316b4a3e34b89309bcb61ef8944b73e56ecd9d9ae87"
     end
     on_intel do
-      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.1/millennium-helpers-v3.0.1-linux-amd64.tar.gz"
+      url "https://github.com/bolens/millenium-helpers/releases/download/v3.0.2/millennium-helpers-v3.0.2-linux-amd64.tar.gz"
       sha256 "5ec3c429c16f4096cb0fb66a561b9fd8f44e8a8fd0d4e4de7cc6ec9a58a53f0c"
     end
   end