+++ go/internal/atomicfile/replace_unix.go
@@ -1,0 +1,9 @@
+//go:build unix
+
+package atomicfile
+
+import "os"
+
+func replaceFile(source, target string) error {
+	return os.Rename(source, target)
+}